Lazy population of HICANN ADC config
To configure the ADCs only when they are accessed (lazy),
we need to keep a copy of the HardwareDatabase passed to Wafer::connect.
However, Wafer::connect takes a const reference to the base class
HardwareDatabase. In the existing scripts, the database is often
constructed as a temporary in the call to Wafer::connect.
Therefore we need to make a copy of the concrete database with the help
of the clone method.

Change-Id: I1a272ac2b733e219af8bbd46765844d731c9d153