Caution fix: keep dlllocking during testing locking
When testing the locking with ReadRepeaterTestdataConfigurator
RepeaterBlocks are read and written to hardware again. Since
dllresetn is write-only the default value in the constructor is
set after reading. This may lead to a wrong value of dllresetb after
writing the read RepeaterBlock. This fix sets dllrestb manually
to 1 to keep the locking. This code should be altered after a
reimplementation of the RepeaterBlock container.

Change-Id: I71dc2bdbc5ea8b27482bd652a6f472d48d449b2e