Make SmartConfigurator (more) thread safe
Deleting or inserting elements into a std::map may lead to undefined
behaviour during concurrent access. Therefore, mWrittenHICANNData
has to be pre-populated such that no new entries have to be
added/deleted during concurrent access. For that:
* a new stage INIT was introduced to the ParallelConfigurators
* the creation of an nullptr entry for every handle was moved to this
  stage

The previous implementation was not sufficient: while the lock ensured
that only one thread inserted new elements in the map at the same time, the
access of the map (e.g. with `map::at`) was still possible concurrently
with the insertion. Since the insertion may lead to a restructuring
of the internal red-black-tree representation of the map,
this lead to occasional `std::out_of_range`-errors.
The addition of the new stage ensures that the map is not accesses
while elements get inserted.

Change-Id: I5bf91193345e69ba796e93ed6044f672d5ee2a01