SynapseController: Only store user configurable data in sthal
Only store data relevant for configuration in sthal, e.g.
content of the configuration register.
Data which does not specify the configuration of the system but rather
an operation the synapse controller should perform to change the
configuration are removed. This for example includes the command in
the control register or the content of the SYNIN register.

Change-Id: Ib0bdd728b8488030066c64cf338f026d09d0bd41